Pass vars previously in CJ77.F90 from driving model rather than hard-code
These variables were switched from being read from config file to being
hard-coded in the initialization subroutine. This limits flexibility for
users of standalone and also of models that embed Cloud-J.

The two physical parameters are not passed but are now hard-coded in
cldj_cmn_mod.F90 (radius of earth and scale heigh above top).

NRANDO is now a locally set variable in cldj_sub_mod.F90 since it is only
used for cloud flag 5 which is not recommended, and should not need to be
changed anyway except by advanced users.

The spherical correction parameters is now set in cldj_cmn_mod.F90 to
2 (refr) since that has been the standard for the last 20 years.

NSBIN is also set as a fixed parameter (27) in cldj_cmn_mod.F90 since it
is only used in Solar-JX.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>